# back/api/swagger/swagger.yaml
swagger: "2.0"
info:
  title: "Ugram API"
  description: "A Ugram API documentation by team 12"
  version: "1.0"
host: "localhost:5000"
basePath: "/"
consumes:
- "application/json"
produces:
- "application/json"
...
schemes:
- "http"
paths:
...
  /follow/followers/count/{id_person}:
    get:
      tags:
      - "follow"
      summary: "Get a person followers number"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Followers number"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
  /follow/followers/name/{id_person}:
    get:
      tags:
      - "follow"
      summary: "Get a person followers names"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Followers names"
          schema:
            $ref: "#/definitions/ResponseObjectPersonNameArray"
  /follow/subscriptions/count/{id_person}:
    get:
      tags:
      - "follow"
      summary: "Get a person subscriptions number"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Subscriptions number"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
  /follow/subscriptions/name/{id_person}:
    get:
      tags:
      - "follow"
      summary: "Get a person subscriptions names"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Subscriptions names"
          schema:
            $ref: "#/definitions/ResponseObjectPersonNameArray"
  /follow/subscriptions/check/{id_person}/{id_person_tocheck}:
    get:
      tags:
      - "follow"
      summary: "Check if a person is subscribed to another"
      parameters:
      - in: "path"
        name: "id_person"
        required: true
        type: "integer"
        description: "Authenticated user"
      - in: "path"
        name: "id_person_tocheck"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Is subscribed : true or false"
          schema:
            $ref: "#/definitions/ResponseObjectBool"
  /follow/subscriptions:
    post:
      tags:
      - "follow"
      summary: "Subscribe to another user"
      parameters:
      - in: "body"
        name: "body"
        description: "Post infos"
        required: true
        schema:
          $ref: "#/definitions/FollowInput"
      responses:
        "201":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
    delete:
      tags:
      - "follow"
      summary: "Unsubscribe"
      parameters:
      - in: "body"
        name: "body"
        description: "Post infos"
        required: true
        schema:
          $ref: "#/definitions/FollowInput"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
...
  /post:
    post:
      tags:
      - "post"
      summary: "Create a new post"
      parameters:
      - in: "body"
        name: "body"
        description: "Post infos"
        required: true
        schema:
          $ref: "#/definitions/PostInputCreate"
      responses:
        "201":
          description: "Id of the newly created post"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
    put:
      tags:
      - "post"
      summary: "Update a post"
      parameters:
      - in: "body"
        name: "body"
        description: "Post infos to update"
        required: true
        schema:
          $ref: "#/definitions/PostInputUpdate"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
    delete:
      tags:
      - "post"
      summary: "Delete a post"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the post to delete"
        required: true
        schema:
          $ref: "#/definitions/PostInputDelete"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
...
  /likepost/count/{id_post}:
    get:
      tags:
      - "likepost"
      summary: "Count likes on a post"
      parameters:
      - in: "path"
        name: "id_post"
        required: true
        type: "integer"
      responses:
        "200":
          description: "The number of likes"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
  /likepost/related/{id_post}:
    get:
      tags:
      - "likepost"
      summary: "Get users who liked the post"
      parameters:
      - in: "path"
        name: "id_post"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Users who liked the post"
          schema:
            $ref: "#/definitions/ResponseObjectPersonNameArray"
  /likepost:
    post:
      tags:
      - "likepost"
      summary: "Create a like on a post"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the post and of the person liking"
        required: true
        schema:
          $ref: "#/definitions/LikepostInput"
      responses:
        "201":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
    delete:
      tags:
      - "likepost"
      summary: "Delete a like on a post"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the post and of the person unliking"
        required: true
        schema:
          $ref: "#/definitions/LikepostInput"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
  /comment/id/{id_comment}:
    get:
      tags:
      - "comment"
      summary: "Get a comment infos"
      parameters:
      - in: "path"
        name: "id_comment"
        required: true
        type: "integer"
      responses:
        "200":
          description: "The comment infos"
          schema:
            $ref: "#/definitions/ResponseObjectComment"
  /comment/post/get/{id_post}:
    get:
      tags:
      - "comment"
      summary: "Get all comments on a post"
      parameters:
      - in: "path"
        name: "id_post"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Comments infos"
          schema:
            $ref: "#/definitions/ResponseObjectCommentArray"
  /comment:
    post:
      tags:
      - "comment"
      summary: "Create a comment"
      parameters:
      - in: "body"
        name: "body"
        description: "Comment infos"
        required: true
        schema:
          $ref: "#/definitions/CommentInputCreate"
      responses:
        "201":
          description: "Id of the newly created comment"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
    delete:
      tags:
      - "comment"
      summary: "Delete a comment"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the post to delete"
        required: true
        schema:
          $ref: "#/definitions/CommentInputDelete"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
  /likecomment/count/{id_comment}:
    get:
      tags:
      - "likecomment"
      summary: "Count likes on a comment"
      parameters:
      - in: "path"
        name: "id_comment"
        required: true
        type: "integer"
      responses:
        "200":
          description: "The number of likes"
          schema:
            $ref: "#/definitions/ResponseObjectInt"
  /likecomment/related/{id_comment}:
    get:
      tags:
      - "likecomment"
      summary: "Get users who liked the comment"
      parameters:
      - in: "path"
        name: "id_comment"
        required: true
        type: "integer"
      responses:
        "200":
          description: "Users who liked the comment"
          schema:
            $ref: "#/definitions/ResponseObjectPersonNameArray"
  /likecomment:
    post:
      tags:
      - "likecomment"
      summary: "Create a like on a comment"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the comment and of the person liking"
        required: true
        schema:
          $ref: "#/definitions/LikecommentInput"
      responses:
        "201":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
    delete:
      tags:
      - "likecomment"
      summary: "Delete a like on a post"
      parameters:
      - in: "body"
        name: "body"
        description: "Id of the comment and of the person unliking"
        required: true
        schema:
          $ref: "#/definitions/LikecommentInput"
      responses:
        "200":
          description: "Empty ResponseObject with potential errors"
          schema:
            $ref: "#/definitions/ResponseObjectEmpty"
...
definitions:
...
  PersonName:
    type: "object"
    properties:
      id_person:
        type: "integer"
        default: "1"
      lastname_person:
        type: "string"
        default: "Doe"
      firstname_person:
        type: "string"
        default: "John"
      username_person:
        type: "string"
        default: "thejohndoe_"
  FollowInput:
    type: "object"
    properties:
      id_person:
        type: "integer"
        default: "1"
      id_person_followed:
        type: "integer"
        default: "2"
  PostInputCreate:
    type: "object"
    properties:
      id_person:
        type: "integer"
        default: "1"
      caption_post:
        type: "string"
        default: "My trip in Italy"
      hashtags_post:
        type: "string"
        default: "#italy"
  PostInputUpdate:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      id_person:
        type: "integer"
        default: "1"
      caption_post:
        type: "string"
        default: "My trip in Italy"
      hashtags_post:
        type: "string"
        default: "#italy"
  PostInputDelete:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      id_person:
        type: "integer"
        default: "1"
        description: "Id of the logged in user"
...
  LikepostInput:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      id_person:
        type: "integer"
        default: "1"
  CommentInputCreate:
    type: "object"
    properties:
      id_post:
        type: "integer"
        default: "1"
      id_person:
        type: "integer"
        default: "1"
      text_comment:
        type: "string"
        default: "..."
  CommentInputDelete:
    type: "object"
    properties:
      id_comment:
        type: "integer"
        default: "1"
  CommentOutput:
    type: "object"
    properties:
      id_comment:
        type: "integer"
        default: "1"
      id_post:
        type: "integer"
        default: "1"
      person:
        $ref: "#/definitions/PersonName"
      isliked_comment:
        type: "boolean"
        default: "true"
      likes_comment:
        type: "integer"
        default: "3"
      text_comment:
        type: "string"
        default: "..."
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
  LikecommentInput:
    type: "object"
    properties:
      id_comment:
        type: "integer"
        default: "1"
      id_person:
        type: "integer"
        default: "1"
...
  ResponseObjectEmpty:
    type: "object"
    properties:
      response:
        type: "object"
        default: null
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
  ResponseObjectBool:
    type: "object"
    properties:
      response:
        type: "boolean"
        default: false
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
  ResponseObjectInt:
    type: "object"
    properties:
      response:
        type: "integer"
        default: 1
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
...
  ResponseObjectPersonNameArray:
    type: "object"
    properties:
      response:
        type: "array"
        items:
          $ref: "#/definitions/PersonName"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
...
  ResponseObjectComment:
    type: "object"
    properties:
      response:
        $ref: "#/definitions/CommentOutput"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
  ResponseObjectCommentArray:
    type: "object"
    properties:
      response:
        type: "array"
        items:
          $ref: "#/definitions/CommentOutput"
      err:
        type: "string"
        enum:
        - "NO"
        - "AUTHENTICATION"
        - "MAIL_EXIST"
        - "USERNAME_EXIST"
        - "OTHER"
        default: "NO"
...